Refactor app's file structure
