+++ src/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 
-import App from "./smart/App.smartcomponent";
+import App from "./app/App.smart";
 
 import { makeStore } from "./store";
 const store = makeStore();
