+++ src/store.js
@@ -1,8 +1,8 @@
 // import { createStore, combineReducers } from "redux";
 import { createStore } from "redux";
 import { combineReducers } from "redux-immutablejs";
-import testi from "./reducers/TestiItem.reducer";
-import turha from "./reducers/Turha.reducer";
+import testi from "./testi-item/TestiItem.reducer";
+import turha from "./app/Turha.reducer";
 
 const combinedReducers = combineReducers({
   testi,
